requirements.txt
fastapi
uvicorn[standard]
torch
diffusers
transformers
accelerate
pillow
python-multipart
